[AAE-6389] improved and moved mocks to proper files
